Replay: merge MessageHandler_* into MsgHandler.h/MsgHandler.cpp
These were trivial classes; having lots of these files around made
it painful to work with.
