+++ Tools/Replay/MsgHandler.cpp
@@ -288,3 +288,283 @@ void MsgHandler::wait_timestamp_from_msg(uint8_t *msg)
     require_field(msg, "TimeMS", timestamp);
     wait_timestamp(timestamp);
 }
+
+
+
+/*
+ * subclasses to handle specific messages below here
+*/
+
+void MsgHandler_AHR2::process_message(uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+    attitude_from_msg(msg, ahr2_attitude, "Roll", "Pitch", "Yaw");
+}
+
+
+void MsgHandler_ARM::process_message(uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+    uint8_t ArmState = require_field_uint8_t(msg, "ArmState");
+    hal.util->set_soft_armed(ArmState);
+    printf("Armed state: %u at %lu\n", 
+           (unsigned)ArmState,
+           (unsigned long)hal.scheduler->millis());
+    dataflash.WriteBlock(msg, f.length);
+}
+
+
+void MsgHandler_ARSP::process_message(uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+
+    airspeed.setHIL(require_field_float(msg, "Airspeed"),
+		    require_field_float(msg, "DiffPress"),
+		    require_field_float(msg, "Temp"));
+    dataflash.WriteBlock(msg, f.length);
+}
+
+
+void MsgHandler_ATT::process_message(uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+    attitude_from_msg(msg, attitude, "Roll", "Pitch", "Yaw");
+}
+
+
+void MsgHandler_BARO::process_message(uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+    baro.setHIL(0,
+		require_field_float(msg, "Press"),
+		require_field_int16_t(msg, "Temp") * 0.01f);
+    dataflash.WriteBlock(msg, f.length);
+}
+
+
+#define DATA_ARMED                          10
+#define DATA_DISARMED                       11
+
+void MsgHandler_Event::process_message(uint8_t *msg)
+{
+    uint8_t id = require_field_uint8_t(msg, "Id");
+    if (id == DATA_ARMED) {
+        hal.util->set_soft_armed(true);
+        printf("Armed at %lu\n", 
+               (unsigned long)hal.scheduler->millis());
+    } else if (id == DATA_DISARMED) {
+        hal.util->set_soft_armed(false);
+        printf("Disarmed at %lu\n", 
+               (unsigned long)hal.scheduler->millis());
+    }
+    dataflash.WriteBlock(msg, f.length);
+}
+
+
+void MsgHandler_GPS2::process_message(uint8_t *msg)
+{
+    // only LOG_GPS_MSG gives us relative altitude.  We still log
+    // the relative altitude when we get a LOG_GPS2_MESSAGE - but
+    // the value we use (probably) comes from the most recent
+    // LOG_GPS_MESSAGE message!
+    update_from_msg_gps(1, msg, false);
+}
+
+
+void MsgHandler_GPS_Base::update_from_msg_gps(uint8_t gps_offset, uint8_t *msg, bool responsible_for_relalt)
+{
+    uint32_t timestamp;
+    require_field(msg, "T", timestamp);
+    wait_timestamp(timestamp);
+
+    Location loc;
+    location_from_msg(msg, loc, "Lat", "Lng", "Alt");
+    Vector3f vel;
+    ground_vel_from_msg(msg, vel, "Spd", "GCrs", "VZ");
+
+    uint8_t status = require_field_uint8_t(msg, "Status");
+    gps.setHIL(gps_offset,
+               (AP_GPS::GPS_Status)status,
+               timestamp,
+               loc,
+               vel,
+               require_field_uint8_t(msg, "NSats"),
+               require_field_uint8_t(msg, "HDop"),
+               require_field_float(msg, "VZ") != 0);
+    if (status == AP_GPS::GPS_OK_FIX_3D && ground_alt_cm == 0) {
+        ground_alt_cm = require_field_int32_t(msg, "Alt");
+    }
+
+    if (responsible_for_relalt) {
+        // this could possibly check for the presence of "RelAlt" label?
+        rel_altitude = 0.01f * require_field_int32_t(msg, "RelAlt");
+    }
+
+    dataflash.WriteBlock(msg, f.length);
+}
+
+
+
+void MsgHandler_GPS::process_message(uint8_t *msg)
+{
+    update_from_msg_gps(0, msg, true);
+}
+
+
+void MsgHandler_IMU2::process_message(uint8_t *msg)
+{
+  update_from_msg_imu(1, msg);
+}
+
+
+void MsgHandler_IMU3::process_message(uint8_t *msg)
+{
+  update_from_msg_imu(2, msg);
+}
+
+
+void MsgHandler_IMU_Base::update_from_msg_imu(uint8_t imu_offset, uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+
+    uint8_t this_imu_mask = 1 << imu_offset;
+
+    if (gyro_mask & this_imu_mask) {
+        Vector3f gyro;
+        require_field(msg, "Gyr", gyro);
+        ins.set_gyro(imu_offset, gyro);
+    }
+    if (accel_mask & this_imu_mask) {
+        Vector3f accel2;
+        require_field(msg, "Acc", accel2);
+        ins.set_accel(imu_offset, accel2);
+    }
+
+    dataflash.WriteBlock(msg, f.length);
+}
+
+
+void MsgHandler_IMU::process_message(uint8_t *msg)
+{
+    update_from_msg_imu(0, msg);
+}
+
+
+void MsgHandler_MAG2::process_message(uint8_t *msg)
+{
+    update_from_msg_compass(1, msg);
+}
+
+
+void MsgHandler_MAG_Base::update_from_msg_compass(uint8_t compass_offset, uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+
+    Vector3f mag;
+    require_field(msg, "Mag", mag);
+    Vector3f mag_offset;
+    require_field(msg, "Ofs", mag_offset);
+
+    compass.setHIL(compass_offset, mag - mag_offset);
+    // compass_offset is which compass we are setting info for;
+    // mag_offset is a vector indicating the compass' calibration...
+    compass.set_offsets(compass_offset, mag_offset);
+
+    dataflash.WriteBlock(msg, f.length);
+}
+
+
+
+void MsgHandler_MAG::process_message(uint8_t *msg)
+{
+    update_from_msg_compass(0, msg);
+}
+
+#include <AP_AHRS.h>
+#include <VehicleType.h>
+
+void MsgHandler_MSG::process_message(uint8_t *msg)
+{
+    const uint8_t msg_text_len = 64;
+    char msg_text[msg_text_len];
+    require_field(msg, "Message", msg_text, msg_text_len);
+
+    if (strncmp(msg_text, "ArduPlane", strlen("ArduPlane")) == 0) {
+	vehicle = VehicleType::VEHICLE_PLANE;
+	::printf("Detected Plane\n");
+	ahrs.set_vehicle_class(AHRS_VEHICLE_FIXED_WING);
+	ahrs.set_fly_forward(true);
+    } else if (strncmp(msg_text, "ArduCopter", strlen("ArduCopter")) == 0 ||
+	       strncmp(msg_text, "APM:Copter", strlen("APM:Copter")) == 0) {
+	vehicle = VehicleType::VEHICLE_COPTER;
+	::printf("Detected Copter\n");
+	ahrs.set_vehicle_class(AHRS_VEHICLE_COPTER);
+	ahrs.set_fly_forward(false);
+    } else if (strncmp(msg_text, "ArduRover", strlen("ArduRover")) == 0) {
+	vehicle = VehicleType::VEHICLE_ROVER;
+	::printf("Detected Rover\n");
+	ahrs.set_vehicle_class(AHRS_VEHICLE_GROUND);
+	ahrs.set_fly_forward(true);
+    }
+    dataflash.Log_Write_Message(msg_text);
+}
+
+
+void MsgHandler_NTUN_Copter::process_message(uint8_t *msg)
+{
+    inavpos = Vector3f(require_field_float(msg, "PosX") * 0.01f,
+		       require_field_float(msg, "PosY") * 0.01f,
+		       0);
+}
+
+
+bool MsgHandler::set_parameter(const char *name, float value)
+{
+    const char *ignore_parms[] = { "GPS_TYPE", "AHRS_EKF_USE" };
+    for (uint8_t i=0; i<sizeof(ignore_parms)/sizeof(ignore_parms[0]); i++) {
+        if (strncmp(name, ignore_parms[i], AP_MAX_NAME_SIZE) == 0) {
+            ::printf("Ignoring set of %s to %f\n", name, value);
+            return true;
+        }
+    }
+    enum ap_var_type var_type;
+    AP_Param *vp = AP_Param::find(name, &var_type);
+    if (vp == NULL) {
+        return false;
+    }
+    if (var_type == AP_PARAM_FLOAT) {
+        ((AP_Float *)vp)->set(value);
+        ::printf("Set %s to %f\n", name, value);
+    } else if (var_type == AP_PARAM_INT32) {
+        ((AP_Int32 *)vp)->set(value);
+        ::printf("Set %s to %d\n", name, (int)value);
+    } else if (var_type == AP_PARAM_INT16) {
+        ((AP_Int16 *)vp)->set(value);
+        ::printf("Set %s to %d\n", name, (int)value);
+    } else if (var_type == AP_PARAM_INT8) {
+        ((AP_Int8 *)vp)->set(value);
+        ::printf("Set %s to %d\n", name, (int)value);
+    } else {
+        // we don't support mavlink set on this parameter
+        return false;
+    }
+    return true;
+}
+
+void MsgHandler_PARM::process_message(uint8_t *msg)
+{
+    const uint8_t parameter_name_len = AP_MAX_NAME_SIZE + 1; // null-term
+    char parameter_name[parameter_name_len];
+
+    require_field(msg, "Name", parameter_name, parameter_name_len);
+
+    set_parameter(parameter_name, require_field_float(msg, "Value"));
+}
+
+
+void MsgHandler_SIM::process_message(uint8_t *msg)
+{
+    wait_timestamp_from_msg(msg);
+    attitude_from_msg(msg, sim_attitude, "Roll", "Pitch", "Yaw");
+}
+
