--- Tools/Replay/MsgHandler_GPS.cpp
@@ -1,6 +1,0 @@
-#include "MsgHandler_GPS.h"
-
-void MsgHandler_GPS::process_message(uint8_t *msg)
-{
-    update_from_msg_gps(0, msg, true);
-}
