--- Tools/Replay/MsgHandler_MAG2.h
@@ -1,11 +1,0 @@
-#include "MsgHandler_MAG_Base.h"
-
-class MsgHandler_MAG2 : public MsgHandler_MAG_Base
-{
-public:
-    MsgHandler_MAG2(log_Format &_f, DataFlash_Class &_dataflash,
-                    uint64_t &_last_timestamp_usec, Compass &_compass)
-        : MsgHandler_MAG_Base(_f, _dataflash, _last_timestamp_usec,_compass) {};
-
-    virtual void process_message(uint8_t *msg);
-};
