--- Tools/Replay/MsgHandler_SIM.cpp
@@ -1,7 +1,0 @@
-#include "MsgHandler_SIM.h"
-
-void MsgHandler_SIM::process_message(uint8_t *msg)
-{
-    wait_timestamp_from_msg(msg);
-    attitude_from_msg(msg, sim_attitude, "Roll", "Pitch", "Yaw");
-}
